Make the code more oop
