--- data/db.txt
@@ -1,1 +1,0 @@
-D | 0 | return book | 2/12/2019 1800
